+++ src/image/tim.cpp
@@ -73,7 +73,31 @@ uint32_t PaletteDetectionStrategyFixed::palOffset(uint16_t, uint16_t) const
 
 uint32_t PaletteDetectionStrategyFixed::palIndex() const
 {
-	return palOffset(0, 0);
+	if (_tim->bpp() == Tim::Bpp16)
+	{
+		return 0;
+	}
+
+	if (_tim->bpp() == Tim::Bpp8 || _tim->paletteWidth() == 16)
+	{
+		return _palY;
+	}
+
+	int palPerLine = _tim->paletteWidth() / 16;
+
+	return _palY * palPerLine + _palX / 16;
+}
+
+bool Tim::save(const char *fileName, bool withAlpha) const
+{
+	PaletteDetectionStrategyFixed fixed(this, 0, 0);
+	return fixed.isValid() && save(fileName, &fixed, withAlpha);
+}
+
+bool Tim::save(const char *fileName, uint8_t paletteId, bool withAlpha) const
+{
+	PaletteDetectionStrategyFixed fixed(this, 0, 0);
+	return fixed.isValid() && save(fileName, &fixed, withAlpha, paletteId);
 }
 
 bool Tim::save(const char *fileName, uint8_t palX, uint8_t palY, bool withAlpha) const
@@ -280,7 +304,7 @@ bool Tim::toRGBA32(uint32_t *target, PaletteDetectionStrategy *paletteDetectionS
 	return true;
 }
 
-bool Tim::save(const char *fileName, PaletteDetectionStrategy *paletteDetectionStrategy, bool withAlpha) const
+bool Tim::save(const char *fileName, PaletteDetectionStrategy *paletteDetectionStrategy, bool withAlpha, int forcePaletteId) const
 {
 	// allocate PBO
 	uint32_t image_data_size = _tim.img_w * _tim.img_h * 4;
@@ -291,7 +315,7 @@ bool Tim::save(const char *fileName, PaletteDetectionStrategy *paletteDetectionS
 	{
 		if (toRGBA32(image_data, paletteDetectionStrategy, withAlpha))
 		{
-			save_texture(image_data, image_data_size, _tim.img_w, _tim.img_h, paletteDetectionStrategy != nullptr ? paletteDetectionStrategy->palIndex() : 0, fileName, false);
+			save_texture(image_data, image_data_size, _tim.img_w, _tim.img_h, forcePaletteId >= 0 ? forcePaletteId : (paletteDetectionStrategy != nullptr ? paletteDetectionStrategy->palIndex() : 0), fileName, false);
 		}
 
 		driver_free(image_data);